Change Sprite when monster is attacking or receiving damage.

+++ Linked/Texture.cpp
@@ -89,7 +89,7 @@ GLuint DynamicTexture::genDynamicGLTexture(int bias)
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST_MIPMAP_NEAREST);
-	glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_LOD_BIAS, bias);
+	glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_LOD_BIAS, (GLfloat)bias);
 	glGenerateMipmap(GL_TEXTURE_2D);
 
 	return textureID;
